crypto-trend-trading/requirements.txt
# Core dependencies
aiohttp>=3.9.0
asyncio>=3.4.3
websockets>=12.0
pyyaml>=6.0.1

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Database
aiosqlite>=0.19.0

# Configuration
pydantic>=2.5.0

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utils
python-dateutil>=2.8.2
